Abstracted ProjectEditor components into separate files
