+++ packages/@codotype/react-components/src/components/project_editor/buildDefaultProject.ts
@@ -1,0 +1,40 @@
+import { GeneratorMeta, Project, ProjectConfiguration } from "./component";
+import "./styles.scss";
+import {
+    ConfigurationGroup,
+    ConfigurationInput,
+    OptionValueInstance,
+    buildConfigurationGroupValue,
+} from "../configuration_group_input";
+
+// // // //
+
+/**
+ * buildDefaultProject
+ * Builds an empty Project
+ * @param
+ */
+export function buildDefaultProject(generatorMeta: GeneratorMeta): Project {
+    // Defines default ProjectConfiguration
+    const projectConfiguration: ProjectConfiguration = generatorMeta.configuration_groups.reduce(
+        (val, configurationGroup: ConfigurationGroup) => {
+            const initialValue: OptionValueInstance = buildConfigurationGroupValue(
+                configurationGroup.properties,
+            );
+            return { ...val, [configurationGroup.identifier]: initialValue };
+        },
+        {}, // Passes in empty ProjectConfiguration
+    );
+
+    // Returns ConfigurationGroupValue
+    const newProject: Project = {
+        label: "My New Project",
+        generatorId: generatorMeta.id,
+        schemas: [],
+        generatorVersion: generatorMeta.version,
+        configuration: projectConfiguration,
+    };
+
+    // Returns the new project
+    return newProject;
+}
